# api-contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateCurrencyConversionRequest" : {
        "properties" : {
          "from_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "from_currency_code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "to_currency_code" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "from_currency_code", "to_currency_code" ],
        "type" : "object"
      },
      "CreateDepositPaymentRouteRequest" : {
        "properties" : {
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_active" : {
            "type" : "boolean"
          },
          "max_amount" : {
            "type" : "number"
          },
          "min_amount" : {
            "type" : "number"
          },
          "note" : {
            "type" : "string"
          },
          "payment_details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "priority" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "deposit_method_id", "title", "treasury_account_public_id" ],
        "type" : "object"
      },
...
      "CurrencyConversionResponse" : {
        "properties" : {
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "from_amount" : {
            "type" : "number"
          },
          "from_currency_code" : {
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rate" : {
            "type" : "number"
          },
          "rate_source" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "to_amount" : {
            "type" : "number"
          },
          "to_currency_code" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DepositPaymentRouteResponse" : {
        "properties" : {
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "deposit_currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "is_active" : {
            "type" : "boolean"
          },
          "max_amount" : {
            "type" : "number"
          },
          "min_amount" : {
            "type" : "number"
          },
          "note" : {
            "type" : "string"
          },
          "payment_details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "priority" : {
            "format" : "int32",
            "type" : "integer"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "treasury_account_code" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_account_title" : {
            "type" : "string"
          },
          "treasury_currency_code" : {
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateDepositPaymentRouteRequest" : {
        "properties" : {
          "is_active" : {
            "type" : "boolean"
          },
          "max_amount" : {
            "type" : "number"
          },
          "min_amount" : {
            "type" : "number"
          },
          "note" : {
            "type" : "string"
          },
          "payment_details" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "priority" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Generated API contract for frontend/backend synchronization.",
    "title" : "Prime Market Backend API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/backoffice/deposit-payment-routes" : {
      "get" : {
        "operationId" : "list_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "deposit_method_id",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "active_only",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DepositPaymentRouteResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-deposit-payment-route-controller" ]
      },
      "post" : {
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDepositPaymentRouteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositPaymentRouteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-deposit-payment-route-controller" ]
      }
    },
    "/api/backoffice/deposit-payment-routes/{publicId}" : {
      "patch" : {
        "operationId" : "update_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDepositPaymentRouteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositPaymentRouteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-deposit-payment-route-controller" ]
      }
    },
...
    "/api/currency-conversions" : {
      "post" : {
        "operationId" : "convert",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCurrencyConversionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CurrencyConversionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "currency-conversion-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Relative to the current backend base URL",
    "url" : "/"
  } ]
}